+++ psutil/_common.py
@@ -178,7 +178,8 @@ sdiskio = namedtuple('sdiskio', ['read_count', 'write_count',
                                  'read_bytes', 'write_bytes',
                                  'read_time', 'write_time'])
 # psutil.disk_partitions()
-sdiskpart = namedtuple('sdiskpart', ['device', 'mountpoint', 'fstype', 'opts'])
+sdiskpart = namedtuple('sdiskpart', ['device', 'mountpoint', 'fstype', 'opts',
+                                     'maxfile', 'maxpath'])
 # psutil.net_io_counters()
 snetio = namedtuple('snetio', ['bytes_sent', 'bytes_recv',
                                'packets_sent', 'packets_recv',
@@ -302,9 +303,6 @@ class NoSuchProcess(Error):
             else:
                 details = "(pid=%s)" % self.pid
             self.msg = "process no longer exists " + details
-
-    def __path__(self):
-        return 'xxx'
 
 
 class ZombieProcess(NoSuchProcess):